Masking Configuration Changes, move logic to lib folder
