+++ lib/kylas_data_mask/masking_configuration.rb
@@ -1,9 +1,7 @@
 # frozen_string_literal: true
 
 module KylasDataMask
-  module ApplicationHelper
-    include KylasDataMask::MaskingConfigurationHelper
-
+  module MaskingConfiguration
     def format_value_based_on_masking(entity_type, user, field_names, value, masking_type)
       field_is_masked?(entity_type, user, field_names) ? masking_based_on_type(value, masking_type) : value
     end
@@ -41,6 +39,17 @@ module KylasDataMask
       Rails.logger.error "KylasDataMask -> masking_based_on_type -> Error Message: #{e.message}"
 
       "****#{value[-3..]}"
+    end
+
+    def cache_masking_configuration(entity_type, user_id, api_key)
+      Rails.cache.fetch("user_#{user_id}_#{entity_type}_mask_configuration", expires_in: 30.minutes) do
+        response = KylasDataMask::MaskedFields.new(api_key: api_key, entity_type: entity_type).fetch
+        response[:data] if response[:success]
+      end
+    end
+
+    def clear_cache_masking_configuration(entity_type, user_id)
+      Rails.cache.delete("user_#{user_id}_#{entity_type}_mask_configuration")
     end
   end
 end
